Set log interval to float so it can be fast in tests

+++ src/prefect/utilities/settings.py
@@ -308,8 +308,8 @@ class LoggingSettings(BaseSettings):
         `{shared_settings.home}/logging.yml`.""",
     )
 
-    orion_log_interval: int = Field(
-        2,
+    orion_log_interval: float = Field(
+        2.0,
         description="""The number of seconds between batched writes of logs to Orion.""",
     )
     orion_max_batch_log_size: int = Field(
